% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gammaFrailty.R
\name{fit_gammaFrailty2}
\alias{fit_gammaFrailty2}
\title{Gamma frailty estimation}
\usage{
fit_gammaFrailty2(
  DATA_LIST = list("DATA", "X"),
  METHOD,
  CPP_CONTROL = list(MAXT = 1000, BURN = 500, STEPSIZE = 0.01, STEPSIZE0 = NULL, NU = 1,
    SEED = 123, PAIRS_RANGE = 5),
  UCMINF_CONTROL = list(ctrl = list(), hessian = 0),
  PAIRS_RANGE,
  STRUCT,
  INIT = NULL,
  ITERATIONS_SUBSET = NULL,
  VERBOSEFLAG = 0
)
}
\arguments{
\item{DATA_LIST}{List containing:
\itemize{
\item{\code{DATA}}{Data matrix with \code{n} rows and \code{p} columns}
\item{\code{X}}{External covariates matrix with \code{n} rows}
}}

\item{METHOD}{Allowed choices are "ucminf", "standard", "bernoulli", "hyper",
"recycle_standard", "recycle_bernoulli", "recycle_hyper" and "randomized".}

\item{CPP_CONTROL}{List of arguments to be passed to the stochastic optimizer:
\itemize{
\item{\code{MAXT}}{Number of iterations.}
\item{\code{BURN}}{Scalar between 0 and 1 denoting the share of \code{n} to be used as burn-in iterations.}
\item{\code{STEPSIZE}}{Initial stepsize parameter.}
\item{\code{NU}}{Nomber of pairs per iteration on average.}
\item{\code{PAR1}}{Hyperparameter for stepsize scheduling by Xu (2011): Scaling.}
\item{\code{PAR2}}{Hyperparameter for stepsize scheduling by Xu (2011): Smallest Hessian eigenvalue.}
\item{\code{PAR3}}{Hyperparameter for stepsize scheduling by Xu (2011): Decay rate}
\item{\code{STEPSIZEFLAG}}{Choose stepsize scheduling: Set 0 for Polyak and Juditsky (1992), 1 for Xu (2011).}
\item{\code{SAMPLING_WINDOW}}{How many iterations to wait before resampling}
\item{\code{EACHCLOCK}}{How often (in terms of iterations) to measure single iteration computational times (using \code{RcppClock})}
\item{\code{SEED}}{Random seed for reproducibility.}
}}

\item{UCMINF_CONTROL}{List of arguments to be passed to \link[ucminf]{ucminf}:
\itemize{
\item{"ctrl"}{Passed to \code{ctrl} argument in \link[ucminf]{ucminf}}
\item{"hessian"}{Passed to \code{hessian} argument in \link[ucminf]{ucminf}}
}}

\item{PAIRS_RANGE}{Maximum lag between pair components.}

\item{STRUCT}{Structure of the correlation matrix. Allowed values are "AR" or "COMPOUND".}

\item{INIT}{Initialization vector.}

\item{ITERATIONS_SUBSET}{Vector containing the indexes of the iterations to be returned to the R session.}

\item{VERBOSEFLAG}{Verbose output.}
}
\description{
Gamma frailty estimation
}
