% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_data}
\alias{generate_data}
\title{Simulate a synthetic dataset}
\usage{
generate_data(INTERCEPT, BETA, X, Q, RHO, SEED, STRUCT)
}
\arguments{
\item{INTERCEPT}{Intercept vector of length \code{p}}

\item{BETA}{Covariates coefficients vector}

\item{X}{External covariates matrix with \code{n} rows}

\item{Q}{Value of q in the data generating process. See Henderson and Shimakura (2003)}

\item{RHO}{Correlation}

\item{SEED}{Seed for reproducibility.}

\item{STRUCT}{Structure of the correlation matrix. Allowed values are "AR" or "COMPOUND".}
}
\description{
Simulate a synthetic dataset
}
