% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tidy_path}
\alias{get_tidy_path}
\title{Get parameters trajectories}
\usage{
get_tidy_path(MOD_OBJ, PATH_LAB, USEREPARTOPAR)
}
\arguments{
\item{MOD_OBJ}{object from \link{fit_gammaFrailty2}}

\item{PATH_LAB}{Label of the path of interest}

\item{USEREPARTOPAR}{TRUE to reparametrize estimate to constrained parameterization.}
}
\description{
Get parameters trajectories
}
