% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partorepar}
\alias{partorepar}
\title{theta reparameterisation}
\usage{
partorepar(par)
}
\arguments{
\item{par}{Constrained parameter vector}
}
\description{
From constrained  to unconstrained parameter vector
}
