% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repartopar}
\alias{repartopar}
\title{theta reparameterisation}
\usage{
repartopar(repar)
}
\arguments{
\item{repar}{Unconstrained parameter vector}
}
\description{
From unconstrained  to constrained parameter vector
}
