% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleVar}
\alias{sampleVar}
\title{Compute sample covariance matrix}
\usage{
sampleVar(
  THETA,
  DATA,
  X,
  STRUCT,
  NU,
  METHOD,
  RANGE,
  TOTFLAG,
  PRINTFLAG,
  PAIRS_RANGE = 100L
)
}
\arguments{
\item{THETA}{Parameter vector}

\item{DATA}{matrix with \code{n} rows and \code{p} columns}

\item{X}{External covariates matrix with \code{n} rows}

\item{STRUCT}{Structure of the correlation matrix. Allowed values are "AR" or "COMPOUND".}

\item{NU}{Nomber of pairs per iteration on average}

\item{METHOD}{Allowed choices are "ucminf", "standard", "bernoulli", "hyper",
"recycle_standard", "recycle_bernoulli", "recycle_hyper"}

\item{RANGE}{Number of iterations after the burn-in.}

\item{TOTFLAG}{Compute total variance}

\item{PRINTFLAG}{Verbose output}

\item{PAIRS_RANGE}{Maximum lag between pair components}
}
\description{
Compute sample covariance matrix
}
